Move fxn to `free_energy.py`
`sample` is not RBFE specific -- uses no definitions from `rbfe.py`, mostly definitions from `free_energy.py`
